Refactor: add comments and update in-memory jsb to remain big endian
This commit introduces better comments for the journaling code so far.

The in memory journal structure's copy of the journal superblock now
remains in big-endian format. This is to follow the convention that
the rest of the ext codebase  uses (the in memory ext superblock's copy
of the on-disk superblock is kept in little endian).

The journal struct types were also updated to uint32_t to align more
with on-disk size.