Add initial logic for revoke structures and functions
Revoke blocks are used to ensure some blocks are not replay else
they would corrupt the filesystem. Two datastructures are used for
tracking revoke blocks since access patterns are different.

First is a linked list backed hashtable. During recovery, the journal
is scanned for revoke blocks and this hashtable of revoke blocks are
built. Any transaction before that revoke block should not replay
the write to that block. This can be a random access pattern so a
hash table is used.

During runtime / normal file operations, if a block is deallocated
a revoke block should be recorded to ensure we write these revoked
/ deallocated blocks to the journal. A linked list is used to track
which blocks are revoked during a transaction.

I also refactored the journal buf struct which probably should have
been a different commit.