Add journal hook to ext2_mkdir
With this new hook, the buffer cache is left in an inconsistent state and panics.
The inconsistent state is most likely an issue with not all file operations
being journaled yet. I might not attempt to debug this and just continue to add
hooks to other file operations as that could just fix it.