Attempt journaling ext2_ind_truncate
The truncate case for shortening a file is a little tricky. First the
inode to shorten should be on the orphan list. The current code tries to
free pointers first then deallocate the resouce pointed. For example,
for block pointers, it will try to free the block pointers then actually
update the bitmap for the block. This sequence is incorrect when
journaling since on replay we need to be able to redo the truncate
process if an inode is on the orphan list. Thus now we update the bitmap
then zero out the blockpointers.